+++ src/drivers/Quartz/Fl_Quartz_Image_Surface_Driver.cxx
@@ -62,7 +62,7 @@ void Fl_Quartz_Image_Surface_Driver::set_current() {
   pre_window = fl_window;
   driver()->gc(offscreen);
   fl_window = 0;
-  ((Fl_Quartz_Graphics_Driver*)driver())->high_resolution( CGBitmapContextGetWidth(offscreen) > width );
+  ((Fl_Quartz_Graphics_Driver*)driver())->high_resolution( CGBitmapContextGetWidth(offscreen) > (size_t)width );
 }
 
 void Fl_Quartz_Image_Surface_Driver::translate(int x, int y) {
